requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.4.2
supabase==2.15.2
python-multipart==0.0.6
gitpython==3.1.40
pygithub==2.1.1